green: AppCoordinator onSave reloads zones after save
Extract watch zone coordinator tests into AppCoordinatorWatchZoneTests
to stay under the 400-line SwiftLint limit.
